Impelement file-based timestream iterators
