+++ pyts2/__init__.py
@@ -4,15 +4,14 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 
-from pyts2.imageio import TSImage
 from pyts2.time import TSInstant
-from pyts2.timestream import TimeStream
+from pyts2.timestream import TimeStream, TimestreamFile
 from pyts2._version import get_versions
 __version__ = get_versions()['version']
 del get_versions
 
 __all__ = [
-    'TSImage',
+    'TimestreamFile',
     'TSInstant',
     'TimeStream',
 ]
